PronunciationApp/requirements.txt
flask
numpy
soundfile
openai-whisper
noisereduce
torch
ffmpeg-python
flask==2.3.2
google-cloud-speech==3.1.2
librosa==0.10.0